+++ client/src/App/App.tsx
@@ -1,0 +1,15 @@
+import React, { useEffect } from 'react';
+import './App.css';
+
+function App() {  
+  useEffect(() => {
+      document.title = "Page not found";
+  });
+  return ( 
+    <div>
+      hi
+    </div>
+  )
+}
+
+export default App;
